refactor: rename core modules and unit tests for conciseness and clarity
