+++ src/wiki_cli/audit.py
@@ -10,9 +10,9 @@ from urllib.parse import unquote
 from rdflib import Graph
 import pyshacl
 
-from .context import WikiConfig, Context
-from .frontmatter import frontmatter_from_path
-from .rdf import frontmatter_to_graph, load_graph
+from .config import WikiConfig, Context
+from .parser import frontmatter_from_path
+from .graph import frontmatter_to_graph, load_graph
 
 logger = logging.getLogger(__name__)
